Replace threadpool with scoped-pool
The scoped-pool create supports both scoped and unscoped threads while
the (more popular) threadpool library only supports unscoped threads.
Unscoped threads are a good use case for executing queries accross
multiple nodes because the query threads continue running even after the
result set has been passed back to the caller. However in the case of
the batch read operation, all threads used to fetch records from
different nodes need to complete before the results are passed back to
the caller. For this use case scoped threads are advantageous because
they can operate on data on the stack in the scope of the calling
thread.