Restrict Travis CI tests to stable/beta/nightly (#84)
Testing against a specific, older Rust version to assure a minimum
supported Rust version for the client is proving futile. Every so often,
Travis CI will pick up an updated dependency (often an indirect
dependency) that requires a newer Rust version, breaking the build. And
because we're following the Cargo team's recommendation to not check in
the Cargo.lock file for library crates there is nothing we can do about
that. So even though the client still compiles fine on older Rust
versions, there is no good way to test that.