[CLIENT-3795] Dropping tokio tasks returns stale data from other commands
Fixes an issue where dropping an async task between the write and the parse stage of a command would cause the connection to be put back in the queue and return stale data from a previous command to an unrelated newer one.

This change closes such connections and throws them away.