WIP: Try to perform data validation
This is not working.  It seems there is some read policy configuration
that I'm not aware of which is impeding my ability to read from the key.
When I manually inspect the results of the read operation, I get _no
results_ at all, where I instead expect to find a bin named binName
which is set to a value "binValue".