WIP: FIX: Code structure improved.
This patch cleans up the vast majority of the build-time warnings when
running cargo test.

This patch also removes the need for clone_safely().  It has come to my
knowledge that the server does properly record the signedness of an
integer when used as map or list payloads.

Finally, this patch re-implements confirm_puts() so that the logic is
easier to follow and verify in code review.

However, as a result of these changes, it now appears that batch Put
operations are not writing to the database.  Note that no errors are
being raised that I can find.  Will continue to research, but I think I
need an extra pair of eyes on this.  :/